Move script.py into client.py
